Refactored frontend code inside root instead of frontend/
