+++ src/components/dashboard/left-navigation/LeftNavigation.vue
@@ -8,6 +8,8 @@
 
 <script>
 
+
+
 import ProfileSection from "@/components/dashboard/left-navigation/ProfileSection.vue";
 import NavigationDropdown from "@/components/dashboard/left-navigation/NavigationDropdown.vue";
