+++ src/components/dashboard/left-navigation/ProfileSection.vue
