+++ src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
 import App from './App.vue'
-import router from "@/router";
+import router from "./router";
 
 const app = createApp(App);
 app.use(router);
